MOS-59 integrate time and timing info with player
I made some modifications to our movie player so that it uses the
time and timing info classes.

This allows the current time of the player to be specified as an
absolute time, which is desirable when we have many sources of data
which different timing information.

As part of this, I provided a new method in the timing info class
to convert an absolute time to an index.

Note that the player is still using QTimer, which only allows
millisecond precision, which is undesirable when we have a frame
rate like 30Hz.

+++ api/isxTrace.h
@@ -21,7 +21,7 @@ public:
     /// \param   start       The start time of the trace.
     /// \param   step        The step time of the trace in milliseconds.
     /// \param   numTimes    The number of time points in the trace.
-    Trace(isx::Time start, isx::Ratio step, uint32_t numTimes);
+    Trace(isx::Time start, isx::Ratio step, uint64_t numTimes);
 
     /// Destructor.
     ///
@@ -31,13 +31,13 @@ public:
     ///
     /// \param   index      The index.
     /// \return             The value at the index.
-    T getValue(uint32_t index) const;
+    T getValue(uint64_t index) const;
 
     /// Write access to a range value by index.
     ///
     /// \param   index      The index.
     /// \param   value      The new value.
-    void setValue(uint32_t index, T value);
+    void setValue(uint64_t index, T value);
 
 private:
