+++ src/isxMovie.cpp
@@ -65,10 +65,7 @@ public:
         // dummy values
         isx::Time start = isx::Time();
         isx::Ratio step(1, 30);
-        // TODO sweet 2016/05/31 : adding this to compile on windows, but
-        // loss of precision may require more thought on design on time or
-        // use of HDF5
-        uint32_t numFrames = static_cast<uint32_t>(m_dims[0]);
+        uint64_t numFrames = m_dims[0];
         m_timingInfo = isx::TimingInfo(start, step, numFrames);
     }
